fix: reenable plugin support
