+++ package.json
@@ -40,7 +40,7 @@
     "@monaco-editor/react": "4.6.0",
     "@terrestris/base-util": "2.0.0",
     "@terrestris/ol-util": "20.0.0",
-    "@terrestris/shogun-util": "9.0.0",
+    "@terrestris/shogun-util": "9.1.0",
     "@uiw/react-md-editor": "4.0.4",
     "antd": "5.20.6",
     "i18next-browser-languagedetector": "8.0.0",
